+++ app/models/dynamicforms/Form.java
@@ -105,7 +105,7 @@ public class Form extends JpaModel {
 			int size = Play.application().configuration().getInt("page.size");
 			if (index < 1)
 				index = 1;
-			Integer rows = (Integer) JPA.em()
+			Long rows = (Long) JPA.em()
 					.createQuery("select count(*) from Form").getSingleResult();
 			List<Form> list = JPA.em()
 					.createQuery("from Form f order by f.name asc")
